requirements.txt
openai>=1.12.0
pinecone-client>=3.0.0,<5.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
PyPDF2>=3.0.0
crewai>=0.28.0
crewai-tools>=0.2.0
streamlit>=1.28.0
